refactor(st-09006): split react node responsibilities
